file sync seems working on real thing

+++ pc_software/my_compare.py
@@ -43,6 +43,9 @@ Files in both but different content: delete from duckypad then write new version
 def delete_path(path):
     if os.path.exists(path) is False:
         return
+    if 'dpp_config.txt' in path:
+        return
+    print("deleting", path)
     if os.path.isfile(path):
         os.remove(path)
     elif os.path.isdir(path):
@@ -60,37 +63,38 @@ def ensure_dir(dir_path):
 
 def copy_file_if_exist(from_path, to_path):
     if os.path.exists(from_path):
+        print(f"\tcopying from {from_path} to {to_path}")
         shutil.copy2(from_path, to_path)
 
 def duckypad_file_sync(sd_dir, modified_dir):
     # top level dirs
     top_level_item_to_add, top_level_item_to_delete, top_level_item_with_difference, common_dirs = compare(sd_dir, modified_dir)
-    print("top_level_item_to_add:", top_level_item_to_add)
-    print("top_level_item_to_delete:", top_level_item_to_delete)
-    print("top_level_item_with_difference:", top_level_item_with_difference)
-    print("common_dirs:", common_dirs)
+    # print("top_level_item_to_add:", top_level_item_to_add)
+    # print("top_level_item_to_delete:", top_level_item_to_delete)
+    # print("top_level_item_with_difference:", top_level_item_with_difference)
+    # print("common_dirs:", common_dirs)
     top_level_to_copy = top_level_item_to_add | top_level_item_with_difference
     top_level_to_remove = top_level_item_to_delete | top_level_to_copy
 
-    print('----------------')
-    print("top_level_to_copy", top_level_to_copy)
-    print("top_level_to_remove", top_level_to_remove)
+    # print('----------------')
+    # print("top_level_to_copy", top_level_to_copy)
+    # print("top_level_to_remove", top_level_to_remove)
 
     for item in top_level_to_remove:
         item_path = os.path.join(sd_dir, item)
-        print("removing...", item_path)
+        # print("removing...", item_path)
         delete_path(item_path)
     
     for item in top_level_to_copy:
         to_copy_source_path = os.path.join(modified_dir, item)
         to_copy_destination_path = os.path.join(sd_dir, item)
-        print("to_copy_source_path:", to_copy_source_path)
-        print("to_copy_destination_path:", to_copy_destination_path)
+        # print("to_copy_source_path:", to_copy_source_path)
+        # print("to_copy_destination_path:", to_copy_destination_path)
         # nothing on top level is worth copying
         if os.path.isfile(to_copy_source_path):
             continue
         # this is a dir, create a new dir first
-        print("creating dir:", to_copy_destination_path)
+        # print("creating dir:", to_copy_destination_path)
         ensure_dir(to_copy_destination_path)
         source_subdir_file_list = [os.path.join(to_copy_source_path, x) for x in os.listdir(to_copy_source_path)]
         source_subdir_file_list = [d for d in source_subdir_file_list if os.path.isfile(d)]
@@ -100,11 +104,11 @@ def duckypad_file_sync(sd_dir, modified_dir):
 
     for common_dir_name in common_dirs:
         files_to_add, files_to_delete, files_with_difference, common_files = compare(os.path.join(sd_dir, common_dir_name), os.path.join(modified_dir, common_dir_name))
-        print('===========', common_dir_name, '===========')
-        print("files_to_add", files_to_add)
-        print("files_to_delete", files_to_delete)
-        print("files_with_difference", files_with_difference)
-        print("common_files", common_files)
+        # print('===========', common_dir_name, '===========')
+        # print("files_to_add", files_to_add)
+        # print("files_to_delete", files_to_delete)
+        # print("files_with_difference", files_with_difference)
+        # print("common_files", common_files)
 
         subdir_file_to_remove = files_to_delete | files_with_difference
         subdir_file_to_remove.add("state.sps")
@@ -114,12 +118,12 @@ def duckypad_file_sync(sd_dir, modified_dir):
             if item.startswith("key"):
                 subdir_file_to_remove.add(item.replace('.txt', '.dsb'))
 
-        print("subdir_file_to_remove:", subdir_file_to_remove)
+        # print("subdir_file_to_remove:", subdir_file_to_remove)
 
         for item in subdir_file_to_remove:
             to_delete_path = os.path.join(sd_dir, common_dir_name)
             to_delete_path = os.path.join(to_delete_path, item)
-            print("\tdeleting...", to_delete_path)
+            # print("\tdeleting...", to_delete_path)
             delete_path(to_delete_path)
 
         subdir_file_to_copy = files_to_add | files_with_difference
@@ -128,21 +132,20 @@ def duckypad_file_sync(sd_dir, modified_dir):
             if item.startswith("key"):
                 subdir_file_to_copy.add(item.replace('.txt', '.dsb'))
 
-        print("subdir_file_to_copy:", subdir_file_to_copy)
+        # print("subdir_file_to_copy:", subdir_file_to_copy)
 
         for item in subdir_file_to_copy:
             copy_from_path = os.path.join(modified_dir, common_dir_name)
             copy_from_path = os.path.join(copy_from_path, item)
             copy_to_path = os.path.join(sd_dir, common_dir_name)
             copy_to_path = os.path.join(copy_to_path, item)
-            print(f"\tcopying from {copy_from_path} to {copy_to_path}")
             copy_file_if_exist(copy_from_path, copy_to_path)
 
-if len(sys.argv) != 3:
-    print(__file__, "sd_dir modified_dir")
-    exit()
+# if len(sys.argv) != 3:
+#     print(__file__, "sd_dir modified_dir")
+#     exit()
 
-sd_dir = sys.argv[1]
-modified_dir = sys.argv[2]
+# sd_dir = sys.argv[1]
+# modified_dir = sys.argv[2]
 
-duckypad_file_sync(sd_dir, modified_dir)
+# duckypad_file_sync(sd_dir, modified_dir)
